requirements.txt
django==4.2.13
djangorestframework==3.15.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
django-environ==0.11.2
redis==5.0.8
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
drf-spectacular==0.27.2
celery==5.4.0
django-redis==5.4.0
sentry-sdk==2.13.0
python-json-logger==2.0.7
cryptography==43.0.1

